重构代码--保存订制师id
