+++ thinkphp/application/index/model/Article.php
@@ -11,20 +11,5 @@ use think\Model;
 
 class Article extends Model
 {
-	/**
-	 * 保存订制师ID 
-	 * @param  id 		$contractorId 订制师ID
-	 * @param  id 		$articleId    文章ID
-	 * @return boolen                 保存成功返回true，否则返回false
-	 */
-    public function saveContractorId($contractorId, $articleId)
-    {
-        $Article = Article::get($articleId);
-        $Article->contractor_id = $contractorId;
-
-        if ($Article->save()) {
-            return true;
-        }
-        return false;
-    }
+	
 }
