+++ thinkphp/application/index/service/Articleservice.php
@@ -185,4 +185,21 @@ class Articleservice
         $Contractor = Contractor::where('article_id',$articleid);
         $Contractor->delete();
     }
+
+    /**
+     * 保存订制师ID 
+     * @param  id       $contractorId 订制师ID
+     * @param  id       $articleId    文章ID
+     * @return boolen                 保存成功返回true，否则返回false
+     */
+    public function saveContractorId($contractorId, $articleId)
+    {
+        $Article = Article::get($articleId);
+        $Article->contractor_id = $contractorId;
+
+        if ($Article->save()) {
+            return true;
+        }
+        return false;
+    }
 }
